[Pytorch Mobile] Improve Bundled Inputs Error Checking
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52386

Remove stale aliasing inputs warning, error check that inputs is not null and has at least one entry, error check that the list of inputs is a list of tuples. This can cause subtle bugs where if the user passes in a list of tensors (the most common mistake) the first dimension of each tensor is dropped. This can go unnoticed because its the often the batch dimension which pytorch occasionally silently re-adds if its missing
ghstack-source-id: 122242626

Differential Revision: [D26374867](https://our.internmc.facebook.com/intern/diff/D26374867/)